requirements.txt
asgiref==3.4.1
astroid==2.8.4
autopep8==1.6.0
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
install==1.3.5
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.4.0
pycodestyle==2.8.0
PyJWT==1.7.1
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.3
